# contest/src/main/java/com/hackathon/contest/payment/model/Payment.java
package com.hackathon.contest.payment.model;

import com.hackathon.contest.fees.model.Fee;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.JoinColumn;
import javax.persistence.ManyToOne;
import java.io.Serializable;
import java.time.Instant;

@Entity
@Builder
@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
public class Payment implements Serializable {
  @Id
  @GeneratedValue(strategy = GenerationType.IDENTITY)
  private Long id;

  private Double amount;

  private Type type;

  @ManyToOne
  @JoinColumn
  private Fee fee;

  private Instant creationDatetime;

  public enum Type {
    CASH, MOBILE_MONEY
  }
}